moved nn to learners directory
