--- medacy/nn/__init__.py
@@ -1,6 +1,0 @@
-"""medacy.nn"""
-from medacy.nn.character_lstm import CharacterLSTM
-from medacy.nn.bilstm_crf import BiLstmCrf
-from medacy.nn.vectorizer import Vectorizer
-from medacy.nn.sequences_dataset import SequencesDataset
-from medacy.nn.bert_crf_for_token_classification import BertCrfForTokenClassification
