+++ medacy/pipeline_components/learners/nn/__init__.py
@@ -1,0 +1,2 @@
+"""medacy.nn"""
+from medacy.pipeline_components.learners.nn.vectorizer import Vectorizer
